fix: 빈 기관 목록 처리 및 count 타입을 Long으로 변경
InstitutionService:
- getActiveInstitutions에 early-return 추가 (빈 목록 시 IN 절 호출 방지)
- count 타입을 Int에서 Long으로 변경하여 오버플로우 방지

Repository 변경:
- AdminRepository.countByInstitutionId: Int → Long
- UserRepository.countByInstitutionId: Int → Long
- UserRepository.countByInstitutionIdAndIsActiveTrue: Int → Long

DTO 변경:
- InstitutionListResponse: adminCount, userCount (Int → Long)
- InstitutionDetailResponse: adminCount, userCount (Int → Long)
- DashboardStatsResponse: totalUsers, activeUsers (Int → Long)
- ChangeInfo: value (Int → Long)
- UserGroupsResponse: totalMembers, ungroupedMembers (Int → Long)

+++ src/main/kotlin/com/onsae/api/admin/repository/AdminRepository.kt
@@ -15,7 +15,7 @@ interface AdminRepository : JpaRepository<Admin, Long> {
     fun existsByInstitutionIdAndEmail(institutionId: Long, email: String): Boolean
     fun findByStatus(status: AccountStatus): List<Admin>
     fun findByInstitutionId(institutionId: Long): List<Admin>
-    fun countByInstitutionId(institutionId: Long): Int
+    fun countByInstitutionId(institutionId: Long): Long
 
     @Query("SELECT a FROM Admin a LEFT JOIN FETCH a.institution LEFT JOIN FETCH a.approvedBy WHERE a.status = :status")
     fun findByStatusWithRelations(status: AccountStatus): List<Admin>
