+++ src/main/kotlin/com/onsae/api/institution/dto/InstitutionListResponse.kt
@@ -24,10 +24,10 @@ data class InstitutionListResponse(
     val directorName: String?,
 
     @Schema(description = "등록된 관리자 수")
-    val adminCount: Int,
+    val adminCount: Long,
 
     @Schema(description = "등록된 사용자 수")
-    val userCount: Int,
+    val userCount: Long,
 
     @Schema(description = "기관 활성화 상태")
     val isActive: Boolean,
